+++ core/src/test/java/com/gentics/mesh/core/project/ProjectEndpointTest.java
@@ -778,8 +778,6 @@ public class ProjectEndpointTest extends AbstractMeshTest implements BasicRestTe
 		String uuid = projectUuid();
 		String branchName = "Branch_V1";
 
-		grantAdmin();
-
 		BranchCreateRequest request = new BranchCreateRequest();
 		request.setName(branchName);
